use host.docker.internal instead of localhost for local pygeoapi
Inside docker containers `host.docker.internal` is the equivalent of `localhost`. River Runner makes an API call to itself when using lat/long to allow for caching of flowpaths. That means for the River Runner process to work, `host.docker.internal` needs to be used as the host url but that breaks any reference to outside of the docker vm. Using `host.docker.internal` instead of `localhost` breaks all intra-pygeoapi links on the web interface.